requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
opencv-python==4.9.0.80
face_recognition==1.3.0
SpeechRecognition==3.10.0
pyttsx3==2.90
psutil==5.9.8
RPi.GPIO==0.7.1
